requirements.txt
playwright==1.44.0
pytest-playwright==0.5.0
python-dotenv==1.0.1